Firebase authentication for movie and user page

+++ src/firebase/Firebase.jsx
@@ -1,13 +1,14 @@
 import { createContext, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
+import { current } from "@reduxjs/toolkit";
 import { initializeApp } from "firebase/app";
 import {getDatabase, onChildAdded, ref, set, onChildRemoved, push} from "firebase/database";
 
 import {updateMovieList, addMovieToList, removeMovieFromList} from "/src/features/movieList/movieListSlice"
-import { updateStreamingServices, setUserName, getUserName } from "../features/userPage/userPageSlice";
+import { setUsername, setUserId, getUserId, addStreamingService, removeStreamingService } from "../features/userPage/userPageSlice";
 import firebaseConfig from "/src/FirebaseConfig.jsx";
 
-import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, onAuthStateChanged } from "@firebase/auth";
+import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, onAuthStateChanged, updateProfile } from "@firebase/auth";
 
 const app = initializeApp(firebaseConfig);
 const auth = getAuth(app);
@@ -18,12 +19,12 @@ let firebase = {
     database: null
 }
 
-export const FirebaseContext = createContext(null);
+export const FirebaseContext = createContext();
 
 export default ({children}) => {
 
     const dispatch = useDispatch();
-    const loggedInUser = useSelector(getUserName);
+    const loggedInUser = useSelector(getUserId);
 
     if (!firebase.api) {
         firebase = {
@@ -48,26 +49,25 @@ export default ({children}) => {
             if (user) {
               // User is signed in, see docs for a list of available properties
               // https://firebase.google.com/docs/reference/js/firebase.User
-              const username = user.uid;
-              console.log("username",loggedInUser)
-              dispatch(setUserName(username));
-              console.log("username",loggedInUser)
-              getMovieList();
-              //getServices();
+              const userId = user.uid;
+              console.log("auth state", userId)
+              console.log("username", user.displayName)
+              dispatch(setUsername(user.displayName))
+              dispatch(setUserId(userId))
               // ...
             } else {
               // User is signed out
               // ...
+              dispatch(setUserId(null))
             }
           });
-    }, [loggedInUser])
+    }, [])
 
-    async function signIn(email, password) {
-        await signInWithEmailAndPassword(auth, email, password)
+    function signIn(email, password) {
+        signInWithEmailAndPassword(auth, email, password)
             .then((userCredential) => {
                 // Signed in 
                 const user = userCredential.user.uid;
-                //dispatch(setUserName(user))
                 console.log(userCredential.user.uid, "logged in")
                 // ...
             })
@@ -78,13 +78,15 @@ export default ({children}) => {
                 // ..
             });
     }
-    async function register(email, password) {
-        await createUserWithEmailAndPassword(auth, email, password)
+    function register(email, password, username) {
+        createUserWithEmailAndPassword(auth, email, password)
             .then((userCredential) => {
                 // Signed in 
                 const user = userCredential.user;
-                dispatch(setUserName(user))
-                console.log(user, "logged in")
+                updateProfile(auth.currentUser, {
+                    displayName: username,
+                })
+                console.log(user, "registered")
                 // ...
             })
             .catch((error) => {
@@ -94,40 +96,43 @@ export default ({children}) => {
                 // ..
             });
     }
-    async function signOutEvent() {
-        await signOut(auth).then(() => {
-            //window.location = "/"
+    function signOutEvent() {
+        signOut(auth).then(() => {
+            window.location = "/"
             // Sign-out successful.
           }).catch((error) => {
             console.log(error)
             // An error happened.
           });
     }
-    
-    //Function for retreiving the movie list from firebase, called from App
-    function getMovieList(){
-        console.log("whoami", loggedInUser)
-        onChildAdded(ref(database, 'movieList/' + loggedInUser), (data) => {
-            console.log("hej", loggedInUser)
-            dispatch(addMovieToList(data.val()))
-        })
-        onChildRemoved(ref(database, 'movieList/' + loggedInUser), (data) => {
-            dispatch(removeMovieFromList(data.val()))
-        })
-    }
 
+    useEffect(() => {
+        if(loggedInUser){
+            onChildAdded(ref(database, 'movieList/' + loggedInUser), (data) => {
+                dispatch(addMovieToList(data.val()))
+            })
+            onChildAdded(ref(database, 'services/' + loggedInUser), (data) => {
+                dispatch(addStreamingService(data.val()))
+            })
+            onChildRemoved(ref(database, 'movieList/' + loggedInUser), (data) => {
+                dispatch(removeMovieFromList(data.val()))
+            })
+            onChildRemoved(ref(database, 'services/' + loggedInUser), (data) => {
+                dispatch(removeStreamingService(data.val()))
+            })
+        }
+    }, [loggedInUser])
+    
     //Functions for adding and removing movies from firebase, called from RenderMovies
-    function addMovie(movie) {
-        console.log("add move user",loggedInUser)
-        const postListRef = ref(database, 'movieList/' + loggedInUser);
-        const postref = push(postListRef);
-        set(postref, movie)
+    function addMovie(movie, userId) {
+        const postListRef = ref(database, 'movieList/' + userId + '/' + movie.imdbId);
+        set(postListRef, movie)
     }
-    function removeMovie(movie) {set(ref(database, 'movieList/' + loggedInUser + '/' + movie.imdbId), null);}
+    function removeMovie(movie, userId) {set(ref(database, 'movieList/' + userId + '/' + movie.imdbId), null);}
 
     //Functions for adding and removing streaming services, called from UserPage
-    function addService(service) {set(ref(database, 'services/' + loggedInUser + service), service);}
-    function removeService(service) {set(ref(database, 'services/' + loggedInUser + service), null);}
+    function addService(service, userId) {set(ref(database, 'services/' + userId + '/' + service), service);}
+    function removeService(service, userId) {set(ref(database, 'services/' + userId + '/' + service), null);}
 
     return (
         <FirebaseContext.Provider value={firebase}>
